requirements.txt
gitpython>=3.1.0
pandas>=2.0.0
pydantic>=2.0.0
xlsxwriter>=3.0.0
